Do not consider paid leaves on weekends

+++ tests/e2e/helpers/timesheetHelper.js
@@ -12,7 +12,7 @@ import managerTeamData from "../data/manager/team.json";
 import managerTaskData from "../data/manager/task.json";
 import { readJSONFile, writeDataToFile } from "../utils/fileUtils";
 import { createProject, deleteProject, getProjectDetails } from "../utils/api/projectRequests";
-import { createTask, deleteTask, likeTask } from "../utils/api/taskRequests";
+import { createTask, deleteTask, likeTask, updateTask } from "../utils/api/taskRequests";
 import { getExchangeRate } from "../utils/api/erpNextRequests";
 import { getEmployeeDetails } from "../utils/api/employeeRequests";
 import { filterApi } from "../utils/api/frappeRequests";
@@ -25,6 +25,7 @@ const empID = process.env.EMP_ID;
 const employeeTimesheetDataFilePath = path.resolve(__dirname, "../data/employee/shared-timesheet.json"); // File path of the employee timesheet data JSON file
 const managerTeamDataFilePath = path.resolve(__dirname, "../data/manager/shared-team.json"); // File path of the manager team data JSON file
 const managerTaskDataFilePath = path.resolve(__dirname, "../data/manager/shared-task.json"); // File path of the manager team data JSON file
+const TASK_TRACKER_PATH = path.resolve(__dirname, "../data/manager/tasks-to-delete.json");
 
 // Global variables to store shared data and reused across functions
 let sharedEmployeeTimesheetData;
@@ -40,7 +41,7 @@ let sharedManagerTaskData;
  * Updates 'payloadCreateTimesheet' and 'payloadFilterTimeEntry' fields with computed dates and employee ID.
  * Saves the updated data back to the shared JSON files.
  *
- * Test Cases: TC2, TC3, TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86
+ * Test Cases: TC2, TC3, TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86, TC96, TC97, TC98, TC99, TC100, TC101
  */
 export const updateTimeEntries = async () => {
   const employeeTimesheetIDs = [
@@ -61,6 +62,12 @@ export const updateTimeEntries = async () => {
     "TC88",
     "TC89",
     "TC90",
+    "TC96",
+    "TC97",
+    "TC98",
+    "TC99",
+    "TC100",
+    "TC101",
   ];
   const managerTeamIDs = ["TC47", "TC49", "TC50"];
 
@@ -106,7 +113,7 @@ export const updateTimeEntries = async () => {
  * This function iterates over predefined time entry payloads and submits them
  * to create timesheet records.
  *
- * Test Cases: TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86
+ * Test Cases: TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86, TC96, TC97, TC98, TC99, TC100, TC101
  */
 export const createTimeEntries = async () => {
   sharedEmployeeTimesheetData = await readJSONFile("../data/employee/shared-timesheet.json");
@@ -127,6 +134,12 @@ export const createTimeEntries = async () => {
     sharedEmployeeTimesheetData.TC88.payloadCreateTimesheet,
     sharedEmployeeTimesheetData.TC89.payloadCreateTimesheet,
     sharedEmployeeTimesheetData.TC90.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC96.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC97.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC98.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC99.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC100.payloadCreateTimesheet,
+    sharedEmployeeTimesheetData.TC101.payloadCreateTimesheet,
     sharedManagerTeamData.TC47.payloadCreateTimesheet,
     sharedManagerTeamData.TC49.payloadCreateTimesheet,
     sharedManagerTeamData.TC50.payloadCreateTimesheet,
@@ -145,7 +158,7 @@ export const createTimeEntries = async () => {
  * This function reads timesheet data from JSON files and iterates through predefined
  * time entry objects, filtering each entry and deleting the corresponding timesheet record.
  *
- * Test Cases: TC2, TC3, TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86
+ * Test Cases: TC2, TC3, TC4, TC5, TC6, TC7, TC14, TC15, TC47, TC49, TC50, TC82, TC83, TC84, TC85, TC86, TC96, TC97, TC98, TC99, TC100, TC101
  */
 export const deleteTimeEntries = async () => {
   sharedEmployeeTimesheetData = await readJSONFile("../data/employee/shared-timesheet.json");
@@ -170,6 +183,12 @@ export const deleteTimeEntries = async () => {
     sharedEmployeeTimesheetData.TC88.payloadFilterTimeEntry,
     sharedEmployeeTimesheetData.TC89.payloadFilterTimeEntry,
     sharedEmployeeTimesheetData.TC90.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC96.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC97.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC98.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC99.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC100.payloadFilterTimeEntry,
+    sharedEmployeeTimesheetData.TC101.payloadFilterTimeEntry,
     sharedManagerTeamData.TC47.payloadFilterTimeEntry,
     sharedManagerTeamData.TC49.payloadFilterTimeEntry,
     sharedManagerTeamData.TC50.payloadFilterTimeEntry,
@@ -242,6 +261,12 @@ export const createProjectForTestCases = async () => {
     "TC88",
     "TC89",
     "TC90",
+    "TC96",
+    "TC97",
+    "TC98",
+    "TC99",
+    "TC100",
+    "TC101",
   ];
   const managerTaskIDs = ["TC22", "TC24", "TC25", "TC26", "TC17", "TC19"];
 
@@ -317,6 +342,12 @@ export const deleteProjects = async () => {
     sharedEmployeeTimesheetData.TC88.payloadDeleteProject.projectId,
     sharedEmployeeTimesheetData.TC89.payloadDeleteProject.projectId,
     sharedEmployeeTimesheetData.TC90.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC96.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC97.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC98.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC99.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC100.payloadDeleteProject.projectId,
+    sharedEmployeeTimesheetData.TC101.payloadDeleteProject.projectId,
     sharedManagerTaskData.TC17.payloadDeleteProject.projectId,
     sharedManagerTaskData.TC19.payloadDeleteProject.projectId,
     sharedManagerTaskData.TC22.payloadDeleteProject.projectId,
@@ -358,6 +389,12 @@ export const createTaskForTestCases = async () => {
     "TC88",
     "TC89",
     "TC90",
+    "TC96",
+    "TC97",
+    "TC98",
+    "TC99",
+    "TC100",
+    "TC101",
   ];
 
   const managerTaskIDs = ["TC22", "TC25", "TC26", "TC17", "TC19"];
@@ -383,6 +420,15 @@ export const createTaskForTestCases = async () => {
 
         // Store task ID in related payloads
         data[testCaseID].payloadDeleteTask.taskID = taskID;
+
+        if (data[testCaseID].payloadUpdateTask) {
+          const updateTaskPayload = data[testCaseID].payloadUpdateTask;
+
+          const updateTaskResponse = await updateTask(taskID, updateTaskPayload);
+          console.log(
+            `UPDATE TASK RESPOSNE FOR ${testCaseID} is : \n ${updateTaskResponse} and custom billable status is ${updateTaskResponse.data.custom_is_billable}`
+          );
+        }
       }
 
       if (data[testCaseID].payloadLikeTask) {
@@ -391,7 +437,7 @@ export const createTaskForTestCases = async () => {
         const response = await likeTask(taskID, data[testCaseID].payloadLikeTask.role);
 
         if (response && typeof response === "object") {
-          console.log(`Successfully liked task for ${testCaseID}:`, response);
+          console.log(`Successfully liked task for ${testCaseID}`);
         } else {
           console.error(`Failed to like task for ${testCaseID}: Unexpected response format`);
         }
@@ -422,21 +468,37 @@ export const createTaskForTestCases = async () => {
 
 /**
  * Deletion of tasks by their name that were created though UI
- */
+ **/
+
 export const deleteByTaskName = async () => {
-  const tasksToBeDeleted = [managerTaskData.TC24.payloadDeleteTaskBySubject.task];
+  try {
+    const tasksToBeDeleted = await readJSONFile(TASK_TRACKER_PATH);
 
-  for (const taskName of tasksToBeDeleted) {
-    console.warn("Checking for task:", taskName);
-
-    const filterResponse = await filterApi("Task", [["Task", "subject", "=", taskName]]);
-    console.warn("Response for getting TASK BY NAME IN DELETION OF TASK IS:", filterResponse);
-
-    if (filterResponse.message?.values?.length) {
-      const taskID = filterResponse.message.values[0];
-      console.log("Task found and ID to delete:", taskID);
-      await deleteTask(taskID);
+    if (!Array.isArray(tasksToBeDeleted) || tasksToBeDeleted.length === 0) {
+      console.warn("No tasks found in the task-tracking file.");
+      return;
     }
+
+    for (const taskName of tasksToBeDeleted) {
+      console.warn("Checking for task:", taskName);
+
+      const filterResponse = await filterApi("Task", [["Task", "subject", "=", taskName]]);
+      console.warn("Response for getting TASK BY NAME IN DELETION OF TASK IS:", filterResponse);
+
+      if (filterResponse.message?.values?.length) {
+        const taskID = filterResponse.message.values[0];
+        console.log("Task found and ID to delete:", taskID);
+        await deleteTask(taskID);
+      } else {
+        console.log(`Task "${taskName}" not found in system. Skipping...`);
+      }
+    }
+
+    //Optionally clear the file after deletion
+    //fs.writeFileSync(TASK_TRACKER_PATH, JSON.stringify([], null, 2), "utf-8");
+    //console.log("Deleted all listed tasks and cleared tracking file.");
+  } catch (error) {
+    console.error("Error while deleting tasks by name:", error.message);
   }
 };
 
@@ -470,6 +532,12 @@ export const deleteTasks = async () => {
     sharedEmployeeTimesheetData.TC88.payloadDeleteTask.taskID,
     sharedEmployeeTimesheetData.TC89.payloadDeleteTask.taskID,
     sharedEmployeeTimesheetData.TC90.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC96.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC97.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC98.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC99.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC100.payloadDeleteTask.taskID,
+    sharedEmployeeTimesheetData.TC101.payloadDeleteTask.taskID,
     sharedManagerTaskData.TC22.payloadDeleteTask.taskID,
     sharedManagerTaskData.TC25.payloadDeleteTask.taskID,
     sharedManagerTaskData.TC26.payloadDeleteTask.taskID,
@@ -557,38 +625,40 @@ export const cleanUpProjects = async (data) => {
 
     // Get Project ID
     const projectRes = await filterApi("Project", [["Project", "project_name", "=", projectName]]);
-    console.warn(`PROJECT RESPONSE for ${key} is ; `, projectRes);
     const projectId = projectRes?.message?.values?.[0]?.[0];
-    console.warn(`Obtained ProjectId value for ${key} is       ; `, projectId);
 
-    if (!projectId) continue;
+    if (!projectId) {
+      continue;
+    } else if (projectId && projectId !== undefined) {
+      console.warn(`\n Obtained ProjectId value for ${key} is       ; `, projectId);
+    }
 
     // Get Task IDs
     const taskRes = await filterApi("Task", [["Task", "project", "=", projectId]]);
-    console.warn(`TASK RESPONSE for ${key} is ; `, taskRes);
 
     const taskRaw = taskRes?.message?.values;
     let taskIds = [];
 
     if (Array.isArray(taskRaw)) {
       taskIds = taskRaw.map((v) => v[0]);
+      console.warn(`OBTAINED TASK for ${key} is: `, taskIds);
     } else {
       console.warn(`No tasks found for project ${key}. Skipping task deletion.`);
     }
 
-    console.warn(`OBTAINED TASK for ${key} is: `, taskIds);
-
     // Get Timesheet IDs
     const timesheetRes = await filterApi("Timesheet", [["Timesheet", "parent_project", "=", projectId]]);
-    console.warn(`TMESHEET RESPONSE for ${key} is ; `, timesheetRes);
+    //console.warn(`TMESHEET RESPONSE for ${key} is ; `, timesheetRes);
 
     const valuesRaw = timesheetRes?.message?.values;
-    console.warn(`Actual values for ${key}:`, valuesRaw);
-    console.warn(`Type of values:`, typeof valuesRaw);
-    console.warn(`Is Array:`, Array.isArray(valuesRaw));
+    //console.warn(`Actual values for ${key}:`, valuesRaw);
+    //console.warn(`Type of values:`, typeof valuesRaw);
+    //console.warn(`Is Array:`, Array.isArray(valuesRaw));
 
     const timesheetValues = timesheetRes?.message?.values || [];
-    console.warn(`OBTAINED TIMESHEET ID FOR ${key} is ; `, timesheetValues);
+    if (Array.isArray(timesheetValues) && timesheetValues.length > 0) {
+      console.warn(`OBTAINED TIMESHEET ID FOR ${key} is ; `, timesheetValues);
+    }
 
     let timesheetIds = [];
 
@@ -597,7 +667,7 @@ export const cleanUpProjects = async (data) => {
       timesheetIds = valuesRaw.flat().filter((v) => typeof v === "string");
     }
 
-    console.warn(`TIMESHEET IDs to delete:`, timesheetIds);
+    //console.warn(`TIMESHEET IDs to delete:`, timesheetIds);
 
     // Store collected info
     deletedData.push({
